Use 16 bit index buffers, since not all Android devices support 32 bit
git-svn-id: svn://srv-dc/svn2/Citrus@209152 20bf7d6a-bd21-9d4d-a3c4-a2c14bb8c68a

+++ Lime/Source/Graphics/Mesh.cs
@@ -31,7 +31,7 @@ namespace Lime
 		public Vector2[] UV2;
 		public Vector2[] UV3;
 		public Vector2[] UV4;
-		public int[] Indices;
+		public ushort[] Indices;
 
 		public Attributes DirtyAttributes;
 		public bool IndicesDirty;
